dark phong but works

+++ src/Texture.hpp
@@ -111,7 +111,7 @@ namespace vogl
     public:
       static Texture *getInstance();
       GLuint addTexture( const std::string& );
-      GLuint addTexture( const glm::uvec3& );
+      GLuint addTexture( const glm::vec3& );
       GLuint getPNGName( const std::string& );
       void makeEnviromentMap( const std::string&, uint );
       GLuint setupEnvMap( const std::string&, uint );
